Added some notes for myself.
Chromosomes will be randomly generated at the start, and assigned to each character.
At each new round, the chromosomes will be mutated/etc and assigned to the new batch (newly spawned) of characters.
Will need to calculate fitness for an individual over time.
Fitness method (in EvolveMain) will take into account the character's fitness instead of the Rand.Next.
Characters will act according to their chromosome.
Do we specify a different trait for each bit/byte in the chromosome or just act as if the chromosome is a fitness value?